counters_0_ver index was not used, fix it so that it is
According to SQLite docs:
"If W is AND-connected terms and X is OR-connected terms and if any term of W
appears as a term of X, then the partial index is usable"

So if my query is "used = 0 AND reserved = 0", then for the index
to be usable the index should be created with "used = 0 OR reserved = 0".
However that condition is true for most entries, so its better
to create an index based on just "used = 0", and store all the reserved values.
Then we get:
	0|0|0|SEARCH TABLE counters USING COVERING INDEX counters_0_ver (used=? AND reserved=? AND ver<?)