Introduce different sizes of dots in redirected sxcp progress output
There are 3 sizes now: short (default,1KB), long (8KB), scaled (equal to blocksize).

Fix bad last line for redirected output. Error was appearing when file
size was equal to number of bytes for each line with dots. Additional
line with summary was printed.

Redirect sxcp output when --dot-size is given
Giving --dot-size to sxcp user wants to use dots output, so enable it without checking if stdout is a tty.