decrement hash refcounts from all distributions
when deleting files during rebalance delete the blocks on the new set

Scenario:
file A on present old volnode set, not yet on the new volnode set
blocks of file A present on old set, and counters already pushed to the new set
counters(old set): 1, counters(new set): 1
sxrm is run, query sent to old volnode set
delete-by-rev queries are sent to both old and new volnodes from the old volnode
  delete-by-rev received by new volnode: file is not present -> delete-by-rev is a no-op here
  delete-by-rev received by old volnode: file is present ->
   check whether we are part of the new volnode set -> no
   check whether we are part of the old volnode set -> yes
     send deletes to hash blocks to the 'dist', i.e. old set
     counters(old set): 0, counters (new set) :1 -- wrong

after this patch
     * send deletes to hash blocks to the new set:
     counters(old set): .. as before, counters(new set): 0 -- correct

Note: the previous code worked correctly in the scenario where
file A's blocks are NOT pushed to the new set yet, and thus deleting
them on the old set prevents them from being replicated in the first place.

Situation is more complicated when the same node is a volnode both on old and
new sets, or if the file is already present on both old and new volnode sets.
In that case we send duplicate delete queries, and rely on the same mechanism
that we rely on during retries on file flush:
 each (replica,fileid,operation) tuple is only applied once, duplicate queries
 are a no-op

P.S. using NL_NEXTPREV or NL_PREVNEXT would be correct too