try to not send negative counters
we can have multiple counters for same (block,replica) with different
rebalance ages (for example: file uploaded then rebalanced then deleted later,
                etc.)
When rebalancing/repairing send the sum of the counters for a given
(block,replica) which should be >= 0 outside of rebalance.
Still if a counter is negative don't overflow it to unsigned.