GC references instead of counters
turn the operations table into a table of revision -> block references
drop the 'used' table that contains the counters
add a table that track the operations on revisions

This is required to support volumes with files that have different replica
counts due to cluster temporarily modified by marking faulty nodes.
It also means that it'll be safe to GC hashes that are shared between
files that were part of a dead replica 1 node, and replica >1 files.

We can GC a block when it has no references.
We can GC a reference (revision) when the sum of its operations is 0.
Keep operations with negative counters (deletes) to avoid race conditions during rebalance.
Rules for GC of reservations is as before: if its TTL expires, or if there was no upload activity on it for <grace_period>.

Make reservation, flush and delete IDs match:
 a reservation is like a +1 except it also has an entry into the reservation table
 the id (hashes) of these operations must match:
  - reserve_id: based on volume/filename so that successfully uploading a file deletes all reservations for it (as before)
  - revision_id: hash of the file's revision and more info to make it unique across the cluster

TODO: upgrade job from 1.0
TODO: allow deleting files even if it has only negative revisions (repeated/replayed deletes)
TODO: optimize file delete: unbump only revision not all hashes
TODO: optimize file flush: bump only revision not all hashes, just do a presence check on the hashes