ensure that PUT/DELETE of same file revision converges to same state on all volnodes (bb920)
File revision creation and deletion can arrive out-of-order on volnode replicas.
If we  observe a delete of a file revision, store a tombstone and skip the
out-of-order file revision creation when it arrives.
Since file revisions are unique we know that by observing a delete the
file revision creation must've happened-before and we need to let the delete
'win'.
This is similar to Observed-Remove Set CRDTs.